+++ easytier/src/web_client/security.rs
@@ -1,12 +1,13 @@
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-use bytes::BytesMut;
 use futures::{SinkExt, StreamExt};
-use snow::{Builder, TransportState, params::NoiseParams};
+use snow::{Builder, params::NoiseParams};
 
 use crate::{
+    common::config::EncryptionAlgorithm,
     proto::common::TunnelInfo,
+    secure_datagram::{SecureDatagramDirection, SecureDatagramSession},
     tunnel::{
         SplitTunnel, StreamItem, Tunnel, TunnelError, ZCPacketSink, ZCPacketStream,
         filter::{TunnelFilter, TunnelWithFilter},
@@ -17,6 +18,9 @@ use crate::{
 const NOISE_MAGIC: &[u8] = b"ET_WEB_NOISE_V1:";
 const NOISE_PROLOGUE: &[u8] = b"easytier-webclient-noise-v1";
 const NOISE_PATTERN: &str = "Noise_NN_25519_ChaChaPoly_SHA256";
+const WEB_SECURE_CIPHER_ALGORITHM: &str = "aes-gcm";
+const WEB_SESSION_GENERATION: u32 = 1;
+const WEB_INITIAL_EPOCH: u32 = 0;
 
 struct RawSplitTunnel {
     info: Option<TunnelInfo>,
@@ -50,24 +54,42 @@ impl Tunnel for RawSplitTunnel {
     }
 }
 
-struct NoiseTunnelFilter {
-    transport: Arc<Mutex<TransportState>>,
+#[derive(Clone, Copy)]
+enum SecureTunnelRole {
+    Initiator,
+    Responder,
 }
 
-impl TunnelFilter for NoiseTunnelFilter {
+impl SecureTunnelRole {
+    fn send_dir(self) -> SecureDatagramDirection {
+        match self {
+            Self::Initiator => SecureDatagramDirection::AToB,
+            Self::Responder => SecureDatagramDirection::BToA,
+        }
+    }
+
+    fn recv_dir(self) -> SecureDatagramDirection {
+        match self {
+            Self::Initiator => SecureDatagramDirection::BToA,
+            Self::Responder => SecureDatagramDirection::AToB,
+        }
+    }
+}
+
+struct SecureDatagramTunnelFilter {
+    session: Arc<SecureDatagramSession>,
+    role: SecureTunnelRole,
+}
+
+impl TunnelFilter for SecureDatagramTunnelFilter {
     type FilterOutput = ();
 
     fn before_send(&self, data: ZCPacket) -> Option<ZCPacket> {
-        let plain = data.tunnel_payload();
-        let mut encrypted = vec![0u8; plain.len() + 64];
-        let len = self
-            .transport
-            .lock()
-            .unwrap()
-            .write_message(plain, &mut encrypted)
-            .ok()?;
-        let mut packet = ZCPacket::new_with_payload(&encrypted[..len]);
+        let mut packet = ZCPacket::new_with_payload(data.tunnel_payload());
         packet.fill_peer_manager_hdr(0, 0, PacketType::Data as u8);
+        self.session
+            .encrypt_payload(self.role.send_dir(), &mut packet)
+            .ok()?;
         Some(packet)
     }
 
@@ -76,23 +98,24 @@ impl TunnelFilter for NoiseTunnelFilter {
             Ok(v) => v,
             Err(e) => return Some(Err(e)),
         };
-        let cipher = packet.payload();
-        let mut plain = vec![0u8; cipher.len() + 64];
-        let len = match self
-            .transport
-            .lock()
+
+        let mut cipher = ZCPacket::new_with_payload(packet.payload());
+        cipher.fill_peer_manager_hdr(0, 0, PacketType::Data as u8);
+        cipher
+            .mut_peer_manager_header()
             .unwrap()
-            .read_message(cipher, &mut plain)
+            .set_encrypted(true);
+        if let Err(e) = self
+            .session
+            .decrypt_payload(self.role.recv_dir(), &mut cipher)
         {
-            Ok(v) => v,
-            Err(e) => {
-                return Some(Err(TunnelError::InvalidPacket(format!(
-                    "noise decrypt failed: {e}"
-                ))));
-            }
-        };
+            return Some(Err(TunnelError::InvalidPacket(format!(
+                "secure datagram decrypt failed: {e}"
+            ))));
+        }
+
         Some(Ok(ZCPacket::new_from_buf(
-            BytesMut::from(&plain[..len]),
+            cipher.payload_bytes(),
             ZCPacketType::DummyTunnel,
         )))
     }
@@ -117,24 +140,50 @@ fn decode_noise_payload(payload: &[u8]) -> Option<&[u8]> {
     payload.strip_prefix(NOISE_MAGIC)
 }
 
+pub fn web_secure_tunnel_supported() -> bool {
+    WEB_SECURE_CIPHER_ALGORITHM
+        .parse::<EncryptionAlgorithm>()
+        .is_ok()
+}
+
+fn web_secure_cipher_algorithm() -> Result<&'static str, TunnelError> {
+    if !web_secure_tunnel_supported() {
+        return Err(TunnelError::InternalError(format!(
+            "web secure tunnel requires {WEB_SECURE_CIPHER_ALGORITHM} support"
+        )));
+    }
+    Ok(WEB_SECURE_CIPHER_ALGORITHM)
+}
+
+fn new_web_secure_session(root_key: [u8; 32], algorithm: &str) -> Arc<SecureDatagramSession> {
+    let algo = algorithm.to_string();
+    Arc::new(SecureDatagramSession::new(
+        root_key,
+        WEB_SESSION_GENERATION,
+        WEB_INITIAL_EPOCH,
+        algo.clone(),
+        algo,
+    ))
+}
+
 fn wrap_secure_tunnel(
     info: Option<TunnelInfo>,
     stream: std::pin::Pin<Box<dyn ZCPacketStream>>,
     sink: std::pin::Pin<Box<dyn ZCPacketSink>>,
-    transport: TransportState,
+    session: Arc<SecureDatagramSession>,
+    role: SecureTunnelRole,
 ) -> Box<dyn Tunnel> {
     let raw = RawSplitTunnel::new(info, stream, sink);
     Box::new(TunnelWithFilter::new(
         raw,
-        NoiseTunnelFilter {
-            transport: Arc::new(Mutex::new(transport)),
-        },
+        SecureDatagramTunnelFilter { session, role },
     ))
 }
 
 pub async fn upgrade_client_tunnel(
     tunnel: Box<dyn Tunnel>,
 ) -> Result<Box<dyn Tunnel>, TunnelError> {
+    let web_cipher_algorithm = web_secure_cipher_algorithm()?;
     let info = tunnel.info();
     let (mut stream, mut sink) = tunnel.split();
 
@@ -159,16 +208,23 @@ pub async fn upgrade_client_tunnel(
     let msg2_packet = stream.next().await.ok_or(TunnelError::Shutdown)??;
     let msg2_cipher = decode_noise_payload(msg2_packet.payload())
         .ok_or_else(|| TunnelError::InvalidPacket("invalid noise msg2 magic".to_string()))?;
-    let mut msg2 = vec![0u8; 1024];
-    state
-        .read_message(msg2_cipher, &mut msg2)
+    let mut root_key_buf = [0u8; 32];
+    let root_key_len = state
+        .read_message(msg2_cipher, &mut root_key_buf)
         .map_err(|e| TunnelError::InvalidPacket(format!("read noise msg2 failed: {e}")))?;
+    if root_key_len != root_key_buf.len() {
+        return Err(TunnelError::InvalidPacket(format!(
+            "invalid web secure root key len: {root_key_len}"
+        )));
+    }
 
-    let transport = state
-        .into_transport_mode()
-        .map_err(|e| TunnelError::InternalError(format!("switch transport mode failed: {e}")))?;
-
-    Ok(wrap_secure_tunnel(info, stream, sink, transport))
+    Ok(wrap_secure_tunnel(
+        info,
+        stream,
+        sink,
+        new_web_secure_session(root_key_buf, web_cipher_algorithm),
+        SecureTunnelRole::Initiator,
+    ))
 }
 
 pub async fn accept_or_upgrade_server_tunnel(
@@ -197,6 +253,7 @@ pub async fn accept_or_upgrade_server_tunnel(
             false,
         ));
     };
+    let web_cipher_algorithm = web_secure_cipher_algorithm()?;
 
     let params: NoiseParams = NOISE_PATTERN
         .parse()
@@ -212,18 +269,54 @@ pub async fn accept_or_upgrade_server_tunnel(
         .read_message(msg1_cipher, &mut msg1)
         .map_err(|e| TunnelError::InvalidPacket(format!("read noise msg1 failed: {e}")))?;
 
+    let root_key = SecureDatagramSession::new_root_key();
     let mut msg2 = vec![0u8; 1024];
     let msg2_len = state
-        .write_message(&[], &mut msg2)
+        .write_message(&root_key, &mut msg2)
         .map_err(|e| TunnelError::InternalError(format!("write noise msg2 failed: {e}")))?;
     sink.send(pack_control_packet(&encode_noise_payload(
         &msg2[..msg2_len],
     )))
     .await?;
 
-    let transport = state
-        .into_transport_mode()
-        .map_err(|e| TunnelError::InternalError(format!("switch transport mode failed: {e}")))?;
+    Ok((
+        wrap_secure_tunnel(
+            info,
+            stream,
+            sink,
+            new_web_secure_session(root_key, web_cipher_algorithm),
+            SecureTunnelRole::Responder,
+        ),
+        true,
+    ))
+}
 
-    Ok((wrap_secure_tunnel(info, stream, sink, transport), true))
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn web_secure_cipher_algorithm_matches_support_flag() {
+        let result = web_secure_cipher_algorithm();
+        if web_secure_tunnel_supported() {
+            assert_eq!(result.unwrap(), WEB_SECURE_CIPHER_ALGORITHM);
+        } else {
+            assert!(matches!(result, Err(TunnelError::InternalError(_))));
+        }
+    }
+
+    #[test]
+    fn web_secure_session_uses_pinned_cipher_algorithm() {
+        if !web_secure_tunnel_supported() {
+            return;
+        }
+
+        let session = new_web_secure_session(
+            SecureDatagramSession::new_root_key(),
+            web_secure_cipher_algorithm().unwrap(),
+        );
+        session
+            .check_encrypt_algo_same(WEB_SECURE_CIPHER_ALGORITHM, WEB_SECURE_CIPHER_ALGORITHM)
+            .unwrap();
+    }
 }
